1. Implemented 'animateMotion', including paced animation (which is still    not supported on other animation elements).  The 'keyPoints' attribute    or the child 'mpath' element are not yet supported, though.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/branches/anim@424731 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/anim/AbstractAnimation.java
@@ -31,10 +31,10 @@ import org.apache.batik.anim.timing.Trace;
 public abstract class AbstractAnimation {
 
     // Constants for calcMode.
-    public final static int CALC_MODE_DISCRETE = 0;
-    public final static int CALC_MODE_LINEAR   = 1;
-    public final static int CALC_MODE_PACED    = 2;
-    public final static int CALC_MODE_SPLINE   = 3;
+    public final static short CALC_MODE_DISCRETE = 0;
+    public final static short CALC_MODE_LINEAR   = 1;
+    public final static short CALC_MODE_PACED    = 2;
+    public final static short CALC_MODE_SPLINE   = 3;
 
     /**
      * The TimedElement that controls the timing of this animation.
